% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_package}
\alias{add_package}
\title{Add R package to renv}
\usage{
add_package(packages, update_snapshot = TRUE)
}
\arguments{
\item{packages}{Character vector of package names}

\item{update_snapshot}{Logical, update renv.lock after installation}
}
\value{
Logical indicating success
}
\description{
Add R package to renv
}
