% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_branch}
\alias{create_branch}
\title{Create feature branch}
\usage{
create_branch(branch_name)
}
\arguments{
\item{branch_name}{Name of the new branch}
}
\value{
Logical indicating success
}
\description{
Create feature branch
}
