% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config_default}
\alias{get_config_default}
\title{Get configuration-aware default value}
\usage{
get_config_default(key, default = NULL)
}
\arguments{
\item{key}{Configuration key}

\item{default}{Default value if config not set}
}
\value{
Configuration value or default
}
\description{
Get configuration-aware default value
}
\keyword{internal}
