% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{git_commit}
\alias{git_commit}
\title{Create and push git commit}
\usage{
git_commit(message, add_all = TRUE)
}
\arguments{
\item{message}{Commit message}

\item{add_all}{Logical, add all files (default: TRUE)}
}
\value{
Logical indicating success
}
\description{
Create and push git commit
}
