% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_push}
\alias{git_push}
\title{Push commits to GitHub}
\usage{
git_push(branch = NULL)
}
\arguments{
\item{branch}{Branch name (default: current branch)}
}
\value{
Logical indicating success
}
\description{
Push commits to GitHub
}
