% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side}

\item{rhs}{Right-hand side}
}
\value{
lhs if not NULL, else rhs
}
\description{
Null-coalescing operator
}
