% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_project}
\alias{join_project}
\title{Join existing zzcollab project (R interface for Developers 2+)}
\usage{
join_project(
  team_name,
  project_name,
  interface = "shell",
  dotfiles_path = NULL,
  dotfiles_nodots = FALSE,
  build_mode = "standard"
)
}
\arguments{
\item{team_name}{Team name (Docker Hub organization)}

\item{project_name}{Project name}

\item{interface}{Interface type: "shell" or "rstudio"}

\item{dotfiles_path}{Path to dotfiles directory}

\item{dotfiles_nodots}{Logical, if TRUE dotfiles need dots added}

\item{build_mode}{Build mode: "fast", "standard", or "comprehensive"}
}
\value{
Logical indicating success
}
\description{
Join existing zzcollab project (R interface for Developers 2+)
}
