% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rebuild}
\alias{rebuild}
\title{Trigger Docker image rebuild}
\usage{
rebuild(target = "docker-build")
}
\arguments{
\item{target}{Build target (default: "docker-build")}
}
\value{
Logical indicating success
}
\description{
Trigger Docker image rebuild
}
