% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{require_makefile}
\alias{require_makefile}
\title{Error if the working directory has no Makefile (not a zzcollab project)}
\usage{
require_makefile()
}
\description{
Error if the working directory has no Makefile (not a zzcollab project)
}
\keyword{internal}
