% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{run_script}
\alias{run_script}
\title{Execute analysis script in container}
\usage{
run_script(script_path, container_cmd = "docker-script")
}
\arguments{
\item{script_path}{Path to R script}

\item{container_cmd}{Make target that runs the script in the container
(default: "docker-script"). The target receives the script via the
SCRIPT make variable.}
}
\value{
Logical indicating success
}
\description{
Execute analysis script in container
}
