% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_project}
\alias{setup_project}
\title{Setup zzcollab project (standard setup, non-init mode)}
\usage{
setup_project(
  dotfiles_path = NULL,
  dotfiles_nodots = FALSE,
  build_mode = "standard",
  base_image = NULL
)
}
\arguments{
\item{dotfiles_path}{Path to dotfiles directory}

\item{dotfiles_nodots}{Logical, if TRUE dotfiles need dots added}

\item{build_mode}{Build mode: "fast", "standard", or "comprehensive"}

\item{base_image}{Base Docker image to use (optional)}
}
\value{
Logical indicating success
}
\description{
Setup zzcollab project (standard setup, non-init mode)
}
