% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_docker_name}
\alias{validate_docker_name}
\title{Validate Docker repository name format}
\usage{
validate_docker_name(name, param_name)
}
\arguments{
\item{name}{Character string to validate}

\item{param_name}{Name of parameter for error messages}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate Docker repository name format
}
\keyword{internal}
