% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_path}
\alias{validate_path}
\title{Validate and normalize file path}
\usage{
validate_path(path, param_name, must_exist = FALSE)
}
\arguments{
\item{path}{Character string path}

\item{param_name}{Name of parameter for error messages}

\item{must_exist}{Logical, whether path must exist}
}
\value{
Normalized path
}
\description{
Validate and normalize file path
}
\keyword{internal}
