% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zzcollab_next_steps}
\alias{zzcollab_next_steps}
\title{Get zzcollab next steps}
\usage{
zzcollab_next_steps()
}
\value{
Character vector with next steps information
}
\description{
Get zzcollab next steps
}
