% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_project}
\alias{setup_project}
\title{Setup zzcollab project (standard setup, non-init mode)}
\usage{
setup_project(base_image = NULL)
}
\arguments{
\item{base_image}{Base Docker image to use (optional)}
}
\value{
Logical indicating success
}
\description{
Setup zzcollab project (standard setup, non-init mode)
}
