% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzcollab-package.R
\docType{package}
\name{zzcollab-package}
\alias{zzcollab}
\alias{zzcollab-package}
\title{zzcollab: Docker-based Research Collaboration Framework}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive framework for reproducible research collaboration using Docker containers. Provides R interfaces to create, manage, and collaborate on research projects with standardized Docker environments, automated CI/CD workflows, and team collaboration tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzcollab}
  \item Report bugs at \url{https://github.com/rgt47/zzcollab/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu} (\href{https://orcid.org/0000-0003-1686-4965}{ORCID})

}
\keyword{internal}
